refactor: extract runtime snapshot helpers
